+++ src/components/CurriculumComponents/CurriculumDownloadView/CurriculumDownloadView.test.tsx
@@ -1,0 +1,27 @@
+// import CurriculumDownloadView from ".";
+
+// import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
+describe("CurriculumDownloadView", () => {
+  test.skip("renders correctly", async () => {
+    // const initialData = {
+    //   school: "test",
+    //   email: "test@example.com",
+    //   downloadType: "word",
+    //   termsAndConditionsSchema: true,
+    // } as const;
+    // const { getByTestId } = renderWithTheme(
+    //   <CurriculumDownloadView data={initialData} />,
+    // );
+    // const schoolElement = getByTestId("download-school");
+    // expect(schoolElement).toHaveValue(initialData.school);
+    // const emailElement = getByTestId("download-email");
+    // expect(emailElement).toHaveValue(initialData.email);
+    // const downloadTypeElement = getByTestId("download-download-type");
+    // expect(downloadTypeElement).toHaveValue(initialData.downloadType);
+    // const termsAndConditionsSchemaElement = getByTestId(
+    //   "download-accept-terms",
+    // );
+    // expect(termsAndConditionsSchemaElement).toBeChecked();
+  });
+});