+++ src/components/CurriculumComponents/CurriculumVisualiser/CurriculumVisualiser.tsx
@@ -240,6 +240,12 @@ const CurriculumVisualiser: FC<CurriculumVisualiserProps> = ({
     setCurrentUnitLessons([]);
   };
 
+  // FIXME: This is kind of a HACK, currently units don't have tier_slug if
+  // they are across multiple year. So we have to do this awkward step.
+  const unitDataTier = unitData?.year
+    ? yearSelection[unitData?.year]?.tier?.tier_slug
+    : undefined;
+
   return (
     <OakGridArea
       id="content"
@@ -254,6 +260,9 @@ const CurriculumVisualiser: FC<CurriculumVisualiserProps> = ({
             const { units, childSubjects, domains, tiers, disciplines } =
               yearData[year] as YearData[string];
 
+            const ref = (element: HTMLDivElement) => {
+              itemEls.current[index] = element;
+            };
             const filteredUnits = units.filter((unit: Unit) =>
               isVisibleUnit(yearSelection, year, unit),
             );
@@ -270,7 +279,7 @@ const CurriculumVisualiser: FC<CurriculumVisualiserProps> = ({
                 $borderRadius={4}
                 className="mobileYearDisplay"
                 id={year}
-                ref={(element) => (itemEls.current[index] = element)}
+                ref={ref}
               >
                 <AnchorTarget
                   $paddingTop={mobileHeaderScrollOffset}
@@ -493,7 +502,11 @@ const CurriculumVisualiser: FC<CurriculumVisualiserProps> = ({
           displayModal={displayModal}
           onClose={handleCloseModal}
           lessons={currentUnitLessons}
-          programmeSlug={createProgrammeSlug(unitData, examboardSlug)}
+          programmeSlug={createProgrammeSlug(
+            unitData,
+            examboardSlug,
+            unitDataTier,
+          )}
           unitOptionsAvailable={unitOptionsAvailable}
           unitSlug={unitData?.slug}
           unitVariantID={unitVariantID}