+++ src/node-lib/curriculum-api-2023/queries/pupilLessonListing/pupilLessonListing.schema.ts
@@ -6,7 +6,6 @@ import { ConvertKeysToCamelCase } from "@/utils/snakeCaseConverter";
 export const lessonBrowseDataSchema = z.array(
   syntheticUnitvariantLessonsSchema.omit({
     null_unitvariant: true,
-    base_slug: true,
   }),
 );